Fora 2.0/requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
Async-OAuthlib==0.0.9
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.23.0
attrs==21.2.0
blinker==1.4
cachetools==5.0.0
canvas==0.3
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
compiler==0.2.0
components==1.2.8
config==0.5.1
coverage==6.2
discord==1.7.3
discord-components==2.1.2
discord-ext-ipc==2.1.1
discord-py-interactions==3.0.2
discord-py-slash-command==3.0.3
discord.py==1.7.3
disrank==0.0.2
distlib==0.3.4
dnspython==2.1.0
easy-pil==0.0.2
env==0.1.0
feedparser==6.0.8
filelock==3.4.2
geographiclib==1.52
geopy==2.2.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.1
ipc==1.2.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
motor==2.5.1
multidict==5.2.0
nest-asyncio==1.5.4
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.1
pluggy==1.0.0
pprintpp==0.4.0
priority==2.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycountry==20.7.3
pycountry-convert==0.7.2
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.6.1
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.3
quart==0.16.2
Quart-Discord==2.1.4
repoze.lru==0.7
requests==2.26.0
sgmllib3k==1.0.0
Shapely==1.5.9
six==1.16.0
tabulate==0.8.9
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
tzwhere==3.0.3
urllib3==1.26.7
virtualenv==20.13.0
websockets==3.4
Werkzeug==2.0.2
wsproto==1.0.0
yarl==1.7.2